Rename team files back to groups
The file renames from PR #13649 (groups → teams) were not reverted when
the subsequent PR was reverted. This renames them back to use "groups"
terminology at the application layer, consistent with the codebase
conventions.

Closes #14034
