+++ apps/prairielearn/src/lib/group-update.sql
