+++ apps/prairielearn/src/lib/submissions-upload.ts
@@ -25,9 +25,9 @@ import {
   type Group,
   RubricItemSchema,
 } from './db-types.js';
+import { createOrAddToGroup, deleteAllGroups } from './groups.js';
 import { type InstanceQuestionScoreInput, updateInstanceQuestionScore } from './manualGrading.js';
 import { createServerJob } from './server-jobs.js';
-import { createOrAddToGroup, deleteAllGroups } from './teams.js';
 
 const sql = sqldb.loadSqlEquiv(import.meta.url);
