+++ apps/prairielearn/src/pages/studentAssessmentInstance/studentAssessmentInstance.ts
@@ -7,8 +7,6 @@ import { loadSqlEquiv, queryRow, queryRows } from '@prairielearn/postgres';
 import * as assessment from '../../lib/assessment.js';
 import { AssessmentInstanceSchema, type File } from '../../lib/db-types.js';
 import { deleteFile, uploadFile } from '../../lib/file-store.js';
-import { idsEqual } from '../../lib/id.js';
-import { type ResLocalsForPage, typedAsyncHandler } from '../../lib/res-locals.js';
 import {
   canUserAssignGroupRoles,
   getGroupConfig,
@@ -16,7 +14,9 @@ import {
   getQuestionGroupPermissions,
   leaveGroup,
   updateGroupRoles,
-} from '../../lib/teams.js';
+} from '../../lib/groups.js';
+import { idsEqual } from '../../lib/id.js';
+import { type ResLocalsForPage, typedAsyncHandler } from '../../lib/res-locals.js';
 import clientFingerprint from '../../middlewares/clientFingerprint.js';
 import logPageView from '../../middlewares/logPageView.js';
 import selectAndAuthzAssessmentInstance from '../../middlewares/selectAndAuthzAssessmentInstance.js';
